wrap @sveltejs/adapter static
This helps us more easily keep up with changes in the official svelte
packages by using it directly and then doing our extra changes
afterward.

One downside to this is that `emptyOutDir` is no longer a valid option,
and the directory is always emptied by the adapter-static adapt
function.

fixes #28